requirements.txt
alembic==1.11.2
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
black==23.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
decorator==5.1.1
Django==4.2.4
dnspython==2.4.2
executing==1.2.0
Faker==19.3.0
idna==3.4
ipdb==0.13.13
ipython==8.14.0
jedi==0.19.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mongoengine==0.27.0
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
rich==13.5.2
six==1.16.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
